Do not clone branches when cloning content
Currently a branch will be cloned if an entity referencing the branch
is cloned. This is a result of the configuration that:

1. Cloning is enabled for content (nodes)
2. Cloning of (related) nodes are enabled by default
3. Cloning of (related) nodes is hidden on the clone form making both
   the problem and the solution harder to figure out.

We cannot configure cloning per bundle/content type so instead we
change the configuration so that cloning of related nodes is disabled
by default.

This could potentially cause us not to clone other nodes that we want
to clone but we will have to cross that bridge if/when we get to it.

We do not change any other setting in this context. The current
configuration that the only clonable entities are nodes, event series
and paragraphs ensures that whatever related cloning settings are set
for other entity types should not matter for us currently.