Add Taxonomy for `Opening Hours` and `Color_Field` Module
We need a taxonomy for different types of Opening Hours. This will be utilized by the API from the custom `dpl_opening_hours` module.

The color_field module can be used in multiple places. Currently, it's needed to add a color field to the term in the taxonomy 'Opening Hours'. This enables users to select a color that will be displayed in the full calendar view.