Replace category exposed filter with facet for events
Views does not seem to support exposed filters at the same time as
facets so replace the current category filter with a facet.

The facet is set up for the events with with the following key
configuration that differ from the default:

- A dropdown widget to replicate the current form element
- Transformation from entity id to label to show category titles
- *Not* hiding facet when source is not rendered to work with AJAX
  views
- Ensure that only one result can be displayed for dropdown widget to
  work

A few other configuration elements are worth mentioning:

- Name/label is set to replicate the current label
- Default option label is set to replicate default Drupal text
- Minimum count set to 1 to hide categories that are not in use for
  events
- Showing the number of results per category

The facet is rendered in the view by adding using a Twig block in the
content list template. For the Events view we can use this to render
the Category facet block.

Provide an updated template for Facets dropdowns which adds classes
and a bit of markup for labels and customized dropdowns.