Initial model for managing opening hours instances and persistance
This adds

- A hook_schema() for creating a database table for storing opening
  hours instances.
- A value object for storing instances with a proper types as we are
  going to get.
- A repository for managing persistance and mapping between database
  rows and value objects.

If mapping fails due to data issues we prefer to handle errors within
the repository, log them and return whatever values we have to the
caller.

Another option for handling opening hours would be to create an entity
for storing each instance, but after some consideration we opted for
a simple table based approach.