Generate code for responding to API requests
We currently use the OpenAPI code generator to generate client code
for calling external APIs based on their provided specification.
However we can also use the same tool to generate code for our own
APIs which will respond to requests. This should ensure that incoming
requests and outgoing responses are handled according to the
specification.

This change:

1. Adds a task for generating a package with the code
2. The .openapi-generator-ignore to ensure we only generate the parts
   we need at the moment
3. Registers the package with Composer

This approach can be extended to other parts of the API in the future.