Add mock image styles, for editors to see card previews. DDFFORM-481
All cards use the same image style, that we then use CSS/Object-Fit
to crop it further, depending on the card "variant".
This means that the image styles would not be visible in the
focal point preview for the editors.
This commit adds the different variants as image styles, although
we do not use them anywhere - it is just so the editors can see
how the image will be cropped, when it is passed through object-fit.