Add new static content submodule with default opening hours categories
This adds five default opening hours categories which should exist
for all libraries when their site is spawned.

We cannot add new entities to existing modules using the default
content module. Consequently we add new submodules which contain new
entities we need to add to existing sites.

The module is hidden to avoid it cluttering up the module list.