Scaffold lagoon.site.yml from project
The default lagoon.site.yml provided by the
amazeeio/drupal_integrations package tries to contact their hosted
version of Lagoon by default.

We need to contact the SSH service in the dplplat01 Lagoon instance.

To achieve this we provide our own version of the lagoon.site.yml
file which contains the host/IP and port to use. To ensure that this
version is used we scaffold our own file on top.