Make sure breadcrumb-url-alias always displays text. DDFFORM-569
Up until now, the token did nothing if it could not find any
breadcrumb context.
This resulted in the branch node not having a URL alias.
We could either create a custom pattern, but I'd rather
just make sure that the breadcrumb token has a solid fallback.
Now, the token will always display the content title, which
works well for branches that have neither breadcrumb structure
or field_categories.