Remove mnsami/composer-custom-directory-installer
This was originally added when adding the Swagger UI module.

This composer installer plugin is no longer needed when placing
named packages in custom directories. Consequently we no longer need
this to place the Swagger UI library in web/libraries.

Consequently we can remove this and reduce complexity in our
dependency chain.