Set danish names for view pages, such as /articles. DDFFORM-502
Ideally, we'd want to actually translate the path, but this
is not an option in Drupal. We could create a URL alias, but then
/articles would still be active.
I found a module named https://www.drupal.org/project/tvp that
seemed to solve the problem, but in reality, it still relies on us
creating a path alias, which has the same problem.