Ensure CMS PHP package does not drift
When we update the specification for the CMS API then our
corresponding PHP package should be updated as well.

We use an specification first approach to the API. The generated
package should help ensure that our request/response handling complies
with the specification.

If the package is not updated with the specification we may face
errors where the specification is updated but the actual code is not.