Make opening hour instance readonly
This should prevent problems down the line caused by mutating states.

This means we have to update the logic for our upsert method a bit as
we cannot return the same instance with the updated id.

This also means that we cannot determine the success based on the
number of affected rows but we were not using that option anyway so
that should not be a loss.