Add initial database support for opening hours repetitions
Update the schema to add a table containing repetitions. By keeping it
in a separate table we avoid duplicating repetition data between
opening hours. We also get a nice incrementing value we can use as id.

For the sake of consistency we also create a repetition for the None
type. This also ensures we have unique ids for these repetitions as
required by our spec.

For now we only handle the None repetition type.

Create an update hook which adds new tables, fields, indices and
values on existing installations.