Add support for handling weekly repetitions
This requires the repository to expand the single provided instance
into multiple instances to be saved.

As creation with weekly repetitions over a period of time can create
many instances we update the specification so that creation returns
an array of opening hours.

This means we have to split up handling of inserts and updates in the
repository.

For the moment we do not support deleting or updating multiple
instances in a repetition. Instead we modify the individual instances.

Expand test coverage to show that it works.