Modfier breadcrumb conditional check.
This check was checking on empty, but the service BreadcrumbHelper is expecting a type of `Drupal\Core\Entity\FieldableEntityInterfac`.

In the case of eventseries, the parameter is a type string, which makes the code break.

This changes the conditional check, to return if the entity is not of that type instead.

DDFFORM-475.