Tidy up admin toolbar, organizing in DDF and Drupal. DDFFORM-617
The toolbar has grown massive, with a lot of useful links.
A lot of the toolbar also promotes stuff that 99% of the editors
will never want to see or use. (Or even have access to).
Instead, we create a DDF dropdown, with relevant settings,
and a Drupal dropdown, for Drupal developers (and admins).