Update max-length of CQL_search_string field.
This is a temporary solution for allowing longer CQL search strings. It is not visually appealing, but as most searchs are copy/pasted into here, it should suffice until the app for enhancing editior work searches is implemented.

This also adds an updatehook to apply the new max-length to the fields already in use.

DDFFORM-494