Reduce access to OpenAPI docs
Currently all authenticated users have access to the OpenAPI Docs.
With updated menu structure this also means that the corresponding
menu item will be visible in the admin UI.

This is a bit confusing as these users do not needs to access the
specification at all.

Update permissions so only relevant roles have access to the API:
- Local administrators and administrators as they may need to document
  things to external partners
- External systems in case they log in to access the administrative UI