Prepare for a takeSnapshot command
Tests should manage how/when to take screenshots individually. We
cannot assume that all our existing tests are suitable for visual
regression testing.

We want to register the takeSnapshot() command but in practice it does
not seem to do anything. To avoid someone trying to use it we
uncomment it for a while.