Patch Drupal Core to provide more details to AJAX form errors
We are setting periodic errors related to AJAX forms which are hard to
track down. Others which so too. Patch Drupal Core to provide
additional information.

Since this is something more people are interested in it should be OK
to patch knowing that we can update the patch once we upgrade to newer
versions of Drupal Core.