Downgrade and pin Drush to 12.1.2
During deployment of release 2024.8.2 we have seen errors in the form:

[warning] Undefined array key "language.en" ConfigImporter.php:353
[warning] Trying to access array offset on value of type null ConfigImporter.php:353
TypeError: array_diff(): Argument #2 must be of type array, null given in array_diff()

This seems to be related to an issue introduced in Drush 12.1.3
https://github.com/drush-ops/drush/issues/5838.
The fix has been merged in the 13.x branch. There is no patch
available which can be applied cleanly. We do not want to upgrade to
a dev release so instead we downgrade to Drush 12.1.2 which according
to the isse is the last version before the bug was introduced.

The fact that this works can be confirmed by recreating the error:

- Download a database from before the problem occurred
- Check out version 2024.8.2 locally
- Run drush deploy which is similar to what occurs when deploying in
  production

The error occurs locally with the current version of Drush, 12.4.3.
It does not occur when downgrading to 12.1.2.

We should upgrade to Drush 13 once this is released.