Prevent failing event deployment hook on new sites
Apparently existing deployment hooks are run on both existing and new
installations. This causes problems with an event deployment hook
where the entity query refers to a field which does not exists on new
installations:

[notice] Deploy hook started: dpl_event_deploy_port_description_fields
[error]  'field_event_description' not found
[error]  Deploy hook failed: dpl_event_deploy_port_description_fields

Update the deployment hook to load all entities but only affect
entities which actually have a nonempty value the source field.

This way the deploy hook can run on new sites but since they do not
have the old field - or it is empty because it is hidden - it will not
do anything.