Return string for update hook feedback instead of using messenger
When using messenger no output is printed in this case for whatever
reason. Having added the message in the messenger bag may result in it
popping up other places.

Ensure that the feedback is printed as a part of running the update
hook by returning it as a string instead.