Enable & setup config_ignore_auto module. DDFFORM-82
We want site administrators to be able to add their own modules, along
with their own settings, and this is not possible with Drupal
configuration management.
In the past, we used `config_ignore`, but this required a massive amount
of work for developers, as all new configuration had to be added
manually.
`config_ignore_auto` works, by us enabling it on production sites, and
then treating any config changes as something that cannot be overriden
in the future by `drush cim`.
See the future ADR for more documentation about pros, cons and effects.