Do the full checks, regardless if AND or OR
Up until now, we only did the full check if we were doing AND, under the
presumption that it was only needed here.
In reality though, we still want to do all the checks, because if a
peice of content has both tags and categories, we want to sort it by
matching tags first, before mixing in categories, branches etc.