Set the default value for branches promoted on lists to false
Without this sorting will be messed up as there is a difference
between a disabled/0 value and an empty/NULL value. Consequently
branches where the value has been toggled will be over branches
where the value has been left untouched.

Branches should either be promoted on top of lists and have a field
value of true/1 or not be promoted and have a field value of false/0.

To ensure this we:

- For future branches we wpdate the configuration of the field to
provide a default value of disabled.
- For current branches we run a deploy hook which sets the value to be
disabled where the field is currently empty.