RelatedContent: Re-factor and share conditiongroup logic
There was some bugs and hacks in the condition group logic, that mostly
had to be fixed by creating seperate condition groups for tags,
categories etc.
This quickly become quite complicated, so I moved it into its own
method, that is able to be called from either getNodeIds() or
getEventInstanceIds().