Remove all content lock restrictions from eventinstances.
Currently event instances can be locked to a user during ediing, similarly to other content types.

This has an interference with recurring_events module. If an editor attempts to change the time for an event series, then for all instances will not work as intended if one or more instances in the series is locked.

Instead of creating a logic for breaking every instances in a series before editing, i suggest we remove all content lock restrictions from instances.

If this evolves into a problem later, we can revert and implement some code to fix this.

DDFFORM-809