Add (temporary?) patch to fix weekly_recurring event start_date determination.
I found an issue in the recurring_events module, which will incorrectly calculate which start date to create  series from. This happens for events of the recur_type weekly_recurring_event, if weekdays are translated into another language than english, since the current comparison does not account for this.

See the description in the added patch for more information.

DDFFORM-809