Migrate existing administrators to local administrators
If anyone due to having excessive access have bumped other users to
a role they should not have access to we want to downgrade them to
the default role given to all administrators.

Webmasters can then in a manual step as a part of the upgrade process
be given the administrator role.