Refactor collation handling
Instead of checking the current state of tables we only focus on the
ones that developers through settings or hooks have specified an
intended state. For these we compare with the current state. If we
have a mismatch then there is something to fix.

The mismatch itself is pulled into a separate value object.

For declaring the intended state through code we add an alter hook
which allows developers to add or modify whatever initial state.

This also means that we no longer have the concept of fallback values.

Update the README accordingly.