Allow authenticated patrons to get opening hours from API
Currently when a user who is authenticated as a patron tries to view
the opening hours for a brach the API request to retrieve the opening
hours is denied with a HTTP 403 response code and message:
"The used authentication method is not allowed on this route."

This happens because the React component by default provides a bearer
token for API requests if it exists. This maps to an authentication
method, dpl_login_user_token, which is currently not enabled for the
REST resource.

This API request does not require any form of authentication but we
cannot easily avoid sending tokens for this request or allow access
as the resource allows anonymous access.

To fix this we add the authentication method to the REST resource.