Update opening hours handling to accept 00:00 as the end time
Currently if an editor enters 00:00 as the end time for an opening
hours instance the API will fail with an error because the API does
not allow end time to occur before start time and 00:00 is considered
to be the first time on a day and not midnight.

To address this we shift the end date one day forward when the time
is 00:00 both when mapping a request to a value object and from
a database record to a value object.

We also update the validation to calculate time difference from start
to end:
- If the difference is inverted then end is before start
- To allow 00:00 the difference must not be greater than a day

Add tests to show that this works both when creating and updating
resources.