Re-introduce missing markup and title for event-list.
As part of moving the automatic paragraphs back to the
pre-RelatedContent state, we re-introduced a view and module.
However, we forgot to introduce the matching twig files,
meaning that the event-list looks different.
The issue is that as part of testing, we only tested a few
results, but the proper styling is only really visible when
several items are added - and the title was not tested.

There is nothing new in this code - it's taken from the previous
state, and it is only temporary.