Make sure nodes are re-saved when placed in the breadcrumb tree. DDFFORM-847
On some node types, we have a structure field, that allows the editor to
place the node as a child of an term.
However, when the actual node is placed in the tree as a parent (as in,
a
term has been created, that links to this node), we want to take this
editor control away, and automatically update the field.
This happens in dpl_breadcrumb_node_presave().
However, if you first setup the node with all it's content, and then go
add it to the tree, no save is triggered on the node, and then it ends
up with a "broken" state.
To fix this, we run `save()` on nodes, when we detect a save on a
breadcrumb term.