Update Adgangsplatformen config to implement DplReactConfigBase
This makes the config element implement CacheableDependencyInterface
which in turn ensures that the proper cache information is used when
adding the configuration as a cacheable dependency in
dpl_react_apps_dpl_react_apps_data().

Without this when passing the object to addCacheableDependency() it
will be treated as an uncacheable dependency which by default will set
the cache max age to 0 and thus make the dynamic cache unusable.

Implementing DplReactConfigBase means that the object has to accept
a ConfigManagerInterface instead of a ConfigFactoryInterface.
Update code and tests accordingly.