Remove all content from dpl_example_breadcrumb.
The module is deprecated and should not be used anymore.

This adds an update hook to uninstall the module if it is installed.

I've left an empty module file, in case the uninstall fails.

The two dependencies ` default_content:default_content ` and `dpl_breadcrumb:dpl_breadcrumb` are for some reason causing a very strange error when performing a `dev:reset`.

For now, they will be kept here. Later when this module is removed this will have to be dealt with.

DDFFORM-921