Refactor current setup for dev and ci setups
The current setup with DOCKER_COMPOSE_FILES variables in Taskfiles
is complex to maintain and makes it difficult to use tasks across
CI and development.

To address this we refactor the variables from Taskfile to a
COMPOSE_FILE variable managed by .env.

We toggle between states using env:* tasks. These are added as
dependencies to existing dev:reset and ci:reset tasks to support
existing workflows.