Update docker compose setup to load database dumps
To avoid using the database dumps in production we add a new
docker-compose.dev.yml file which adds and modifies services to use
the dumps.

Update our environment handling to load this new file and drop the
corresponding steps from the reset task.

We currently use it both for CI and development as the reset script
for both situations no longer installs Drupal on its own.