Fix template for filtered event list
After reintroducing related content filtered event list are rendered
without any content. This is because the reintroduction changed
variable names and there is no variable view_has_results which matches
the current template.

The @todo in the template marks the template for deletion after
reintroducing related content. However this does not work as expected.
It will render the list but also the name of the paragraph.

Consequently to address this we only render the view. This makes it
work the same way as
cms/web/themes/custom/novel/templates/paragraphs/paragraph--card-grid-automatic.html.twig