Update removal of default form actions on eventseries confirm page
When updating the schedule for an event series the recurring events
module shows a summary of the change and asks the user to confirm it
as this will recreate all instances.

In this situation we want to hide the default form actions as they
make little sense in this situation even though they are fully
functional.

Until now we have achieved this by removing the form elements from the
form.

After updating the Gin administration theme and Drupal Core from 10.1
to 10.3 this approach no longer works. Gin expects the actions to be
available and throws a hard PHP error if they are not.

Attempt to address this using alternate approaches using the Drupal
Form API e.g. #access=FALSE or #disabled=TRUE has not worked due to
the rather complex form manipulations of Gin.

Instead we switch to a CSS-based approach which hides the form
elements instead. This should not make a huge difference compared to
the original approach.

Gin also does interesting things to the DOM structure so the resulting
CSS selector also becomes a bit complex.