Re-delete translation of webform.settings.yml, & remove faulty `zip` key
Apparantly, the faulty keys in the webform.settings.yml that get
translated and duplicated as 2x"Postnummer" still exists - our hope
was that it was some kind of broken state.
Instead, we'll remove the key from the source config, so it does
not get translated at all.
Apparantly, the data is test data, so it should be safe.

After doing this, we'll also re-delete the translation, by just
bumping the hook_update number