Delete faulty local string translation of webform.settings
It turns out the way configuration translation is handled by the
config_translation module is:

1. Import the translation strings from the remove source into Drupal translation
   strings.
2. Import translation strings into configuration.

This causes problems when we want to remove a local faulty translation as in
dpl_webform_update_10003(). Even though we have removed the translated
configuration (1) it still lingers as a translation string (2). This will be i
ncluded in the translated configuration when rerunning the configuration
translation process.

Consequently: To remove the faulty translation for webform.settings entirely
we have to remove it from the local string translations as well.

The problematic part is in the test names part of the configuration so we can
query based on this.