Add initial cypress test for testing events
This test creates an event series with a single instance and asserts
that it has been created correctly.

The test is implemented with a cleanup step before hand as this is
the recommended approach by Cypress.
https://docs.cypress.io/guides/references/best-practices#State-reset-should-go-before-each-test

In the implementation we have to use force for selecting the
recur type. This is the suggested approach when using Select2.
https://www.cypress.io/blog/2020/03/20/working-with-select-elements-and-select2-widgets-in-cypress