Disable the late runtime processor when using CLI
When running drush deploy on production environments we are frequently
seeing problems where our process fails with the following error:

Drupal\Core\Database\DatabaseExceptionWrapper: SQLSTATE[HY000]:
General error: 2006 MySQL server has gone away: SELECT * FROM
"purge_queue" q WHERE ((expire = 0) OR (:now > expire)) ORDER BY
created, item_id ASC LIMIT 0, 100;

It occurs for little over 10% of production deployments. Redeploying
the site will make the problem go away.

This query occurs because the late runtime processor tries to claim
items from the purge queue at the end of Drush commands.

We have not been able to identify the root cause of this issue. Based
on the metrics and logs we have available the database does not seem
to be under significant strain. There should not be a large quantity
of items in the queue etc.

To avoid this from happening we have discussed alternate solutions:

- Reducing the number of items claimed does not make a difference. We
tried going from 100 to 30.
- We could move the purge queue to alternate storage. Here Redis comes
to mind since it is already used in the project. However it is worth
noting that this would require Redis with persistence to avoid loosing
items if Redis restarts or is moved within the cluster. Our current
use of Redis is not persistent and this would also require introducing
a sandbox module to support Redis with Purge.
https://www.drupal.org/sandbox/marvin_b8/3417595

In the end we go with a simpler solution: Not implicitly processing
the purge queue when running CLI commands. This way we also avoid
it during deployment Drush commands which is where this error occurs.

For deployments this should not make a difference. We already
explictly purge the entire Varnish Cache when doing a deployment.
https://github.com/danskernesdigitalebibliotek/dpl-platform/blob/main/infrastructure/dpladm/env-repo-template/standard/.lagoon.yml#L31

There might be other situations where an update which happens during
cron invoked with drush cli would place items in the purge queue. Here
these items would not be processed immediately. They will have to wait
for the next event which will trigger purge queue processing e.g.
an uncached web request or a drush cron.

We expect both events to occur frequently enough that this should not
be a problem.