Add health check module
This module provides an uncached endpoint which checks whether Drupal
is able to bootstrap.

By checking this we should ensure that the site is in a relatively
healthy state - most importantly that we have a connection to the
database and that the cache is not corrupt.

The health check module provides a relatively simple module for doing
so, it is well-maintained and we have good experience using it in
other projects.