Support cookie-based authentication for campaign endpoint
This will allow editors test their campaigns in the same browser as
they are creating the campaign in. Without this they would have to log
out and test the campaign in an anonymous browser.

The situation is caused by the fact that the request contains a cookie
which is an authentication method, but the campaign endpoint is not
configured to support it.

To address this we add support for cookie-based authentication. For
this to work out of the box we need to disaable the X-CSRF requirement
which we have also done elsewhere. For this endpoint it should not
make a difference as it is usable for all types of users and thus does
not need extra protecton.

Update the specification and generated packages accordingly.