Add a link to editorial search on material search page. DDFFORM-678
When on the material search, we can look up the editorial (web) search
using the same fulltext query, and see if there are any results.
If there are results, we add a link.
This link should also be parsed along to the react app.
This PR prepares that, sending it along as `web-search-link-text` and
`web-search-link-url`.
At the time of writing this, the react app has *not* been updated to
take these values yet.