Add 1 hour time and tag based cache for events view
Currently events views do not use caching. This is because we usually
do not want to show events which have occurred (and are thus less
relevant) *and* we want editorial changes to take effect as quickly
as possible.

Handling the relative time of event occurrence and the editorial
changes (which is usually related to cache tags) has not been
possible.

To address this situation we have added a patch for Search API views.
This adds a cache plugin which supports both.

Configure this new cache plugin for our event based views with a
maximum of 1 hour cache lifetime.

This should ensure that the event list gets cached to reduce resource
usage and produce faster responses and that cached results are updated
relatively frequently so we do not show outdated responses even if
there are no editorial changes.

Results should be cached both in Varnish which targets anonymous users
and in the Drupal Dynamic Cache which targets patrons and editorial
users.