Reduce load from crawlers by requesting 10s delay between requests
We have seen sites being overloaded by requests from web crawlers.
A report on user agents have revealed GoogleBot, SEMrushbot etc.

To reduce the load these might cause on the system we append a request
for a 10s delay between requests in robots.txt.

Although it is up to the individual bot whether to adhere to this,
some bots do e.g.
- https://help.ahrefs.com/en/articles/78158-how-do-i-control-your-bot-s-crawling-behaviour
- https://www.semrush.com/kb/1056-optimize-site-audit-crawl-speed

Other bots - notably GoogleBot - do not.

Using a 10s delay is a somewhat arbitrary value but it is the value
suggested by https://search.gov/indexing/robotstxt.html which seems
like a relatable source. It is also the maximum value respected by
SEMrushbot.

We could add an include- or exclude-list for supported crawlers.
However that seems excessive at the moment as we do not know what
external services an individual library site may want to use and the
setting is not currently configurable.